Start refactoring packages, temaples and i18n.porps.
