+++ src/main/java/org/everit/jira/timetracker/plugin/util/JiraTimetrackerUtil.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.everit.jira.timetracker.plugin;
+package org.everit.jira.timetracker.plugin.util;
 
 import java.util.Date;
 import java.util.List;
