+++ src/main/resources/atlassian-plugin.xml
@@ -132,57 +132,57 @@
 
   <webwork1 key="jttp-webaction" name="JTTP Webwork actions" class="java.lang.Object">
     <resource type="i18n" name="i18n-jttp-vm" location="i18n/jttp" />
-    <resource type="i18n" name="i18n-jttp_feedback-vm" location="i18n/jttp_feedback" />
+    <resource type="i18n" name="i18n-feedback-vm" location="i18n/feedback" />
     <actions>
-      <action name="org.everit.jira.timetracker.plugin.JiraTimetrackerWebAction" alias="JiraTimetrackerWebAction">
-        <view name="success">/templates/jttp.vm</view>
-        <view name="input">/templates/jttp.vm</view>
-        <view name="error">/templates/error.vm</view>
+      <action name="org.everit.jira.timetracker.plugin.web.JiraTimetrackerWebAction" alias="JiraTimetrackerWebAction">
+        <view name="success">/templates/timetracker/jttp.vm</view>
+        <view name="input">/templates/timetracker/jttp.vm</view>
+        <view name="error">/templates/common/error.vm</view>
       </action>
-      <action name="org.everit.jira.timetracker.plugin.JiraTimetrackerWorklogsWebAction" alias="JiraTimetrackerWorklogsWebAction">
-        <view name="success">/templates/missing_worklogs_report.vm</view>
-        <view name="input">/templates/missing_worklogs_report.vm</view>
-        <view name="error">/templates/error.vm</view>
-      </action>
-
-      <action name="org.everit.jira.timetracker.plugin.JiraTimetrackerChartWebAction" alias="JiraTimetrackerChartWebAction">
-        <view name="success">/templates/chart_report.vm</view>
-        <view name="input">/templates/chart_report.vm</view>
-        <view name="error">/templates/error.vm</view>
+      <action name="org.everit.jira.timetracker.plugin.web.JiraTimetrackerWorklogsWebAction" alias="JiraTimetrackerWorklogsWebAction">
+        <view name="success">/templates/timetracker/missing_worklogs_report.vm</view>
+        <view name="input">/templates/timetracker/missing_worklogs_report.vm</view>
+        <view name="error">/templates/common/error.vm</view>
       </action>
 
-      <action name="org.everit.jira.timetracker.plugin.JiraTimetrackerSettingsWebAction" alias="JiraTimetrackerSettingsWebAction">
-        <view name="success">/templates/settings.vm</view>
-        <view name="input">/templates/settings.vm</view>
-        <view name="error">/templates/error.vm</view>
+      <action name="org.everit.jira.timetracker.plugin.web.JiraTimetrackerChartWebAction" alias="JiraTimetrackerChartWebAction">
+        <view name="success">/templates/timetracker/chart_report.vm</view>
+        <view name="input">/templates/timetracker/chart_report.vm</view>
+        <view name="error">/templates/common/error.vm</view>
       </action>
-      <action name="org.everit.jira.timetracker.plugin.AdminSettingsWebAction" alias="JiraTimetrackerAdminSettingsWebAction">
-        <view name="success">/templates/adminsettings.vm</view>
-        <view name="input">/templates/adminsettings.vm</view>
-        <view name="error">/templates/error.vm</view>
+
+      <action name="org.everit.jira.timetracker.plugin.web.JiraTimetrackerSettingsWebAction" alias="JiraTimetrackerSettingsWebAction">
+        <view name="success">/templates/timetracker/settings.vm</view>
+        <view name="input">/templates/timetracker/settings.vm</view>
+        <view name="error">/templates/common/error.vm</view>
+      </action>
+      <action name="org.everit.jira.timetracker.plugin.web.AdminSettingsWebAction" alias="JiraTimetrackerAdminSettingsWebAction">
+        <view name="success">/templates/timetracker/adminsettings.vm</view>
+        <view name="input">/templates/timetracker/adminsettings.vm</view>
+        <view name="error">/templates/common/error.vm</view>
       </action>
       
-      <action name="org.everit.jira.timetracker.plugin.JiraTimetrackerTableWebAction" alias="JiraTimetrackerTableWebAction">
-        <view name="success">/templates/table_report.vm</view>
-        <view name="input">/templates/table_report.vm</view>
-        <view name="error">/templates/error.vm</view>
+      <action name="org.everit.jira.timetracker.plugin.web.JiraTimetrackerTableWebAction" alias="JiraTimetrackerTableWebAction">
+        <view name="success">/templates/timetracker/table_report.vm</view>
+        <view name="input">/templates/timetracker/table_report.vm</view>
+        <view name="error">/templates/common/error.vm</view>
       </action>
-      <action name="org.everit.jira.timetracker.plugin.JiraReportingPopupWebAction" alias="JiraReportingPopupWebAction">
-        <view name="success">/templates/reporting_popup.vm</view>
-        <view name="input">/templates/reporting_popup.vm</view>
-        <view name="error">/templates/error.vm</view>
+      <action name="org.everit.jira.timetracker.plugin.web.JiraReportingPopupWebAction" alias="JiraReportingPopupWebAction">
+        <view name="success">/templates/reporting/reporting_popup.vm</view>
+        <view name="input">/templates/reporting/reporting_popup.vm</view>
+        <view name="error">/templates/common/error.vm</view>
       </action>
     </actions>
   </webwork1>
   
   <webwork1 key="jtrp-webaction" name="JTRP Webwork actions" class="java.lang.Object">
     <resource type="i18n" name="i18n-jtrp-vm" location="i18n/jtrp" />
-    <resource type="i18n" name="i18n-jttp_feedback-vm" location="i18n/jttp_feedback" />
+    <resource type="i18n" name="i18n-feedback-vm" location="i18n/feedback" />
     <actions>
-      <action name="org.everit.jira.timetracker.plugin.ReportingSettingsWebAction" alias="JiraTimetrackerReportingSettingsWebAction">
-        <view name="success">/templates/reportingsettings.vm</view>
-        <view name="input">/templates/reportingsettings.vm</view>
-        <view name="error">/templates/error.vm</view>
+      <action name="org.everit.jira.timetracker.plugin.web.ReportingSettingsWebAction" alias="JiraTimetrackerReportingSettingsWebAction">
+        <view name="success">/templates/reporting/reportingsettings.vm</view>
+        <view name="input">/templates/reporting/reportingsettings.vm</view>
+        <view name="error">/templates/common/error.vm</view>
       </action>
     </actions>
   </webwork1>
