+++ src/main/resources/templates/common/analytics.vm
