+++ src/main/resources/templates/feedback/feedback_table.vm
@@ -22,7 +22,7 @@
         <h1 class="aui-dialog2-header-main" style="color: #707070;">$i18n.getText('jttp.feedback.header')</h1>   
       </header>
       <!-- Main dialog content -->
-      #parse("templates/feedback_content.vm")
+      #parse("templates/feedback/feedback_content.vm")
       <!-- Dialog footer -->
       <footer class="aui-dialog2-footer">
         <!-- Actions to render on the right of the footer -->
