+++ src/main/resources/templates/reporting/reporting_popup.vm
