+++ src/main/resources/templates/timetracker/summary.vm
