+++ src/main/resources/templates/timetracker/table_report.vm
@@ -15,26 +15,23 @@
  *#
 <html>
 <head>
-<title>Timetracker Chart Report</title>
-
+<title>Timetracker Table Report</title>
 </head>
 $webResourceManager.requireResource("jira.webresources:global-static")
 $webResourceManager.requireResource("jira.webresources:calendar")
 $webResourceManager.requireResource("jira.webresources:calendar-en")
 $webResourceManager.requireResource("jira.webresources:dwr", null)   
 $webResourceManager.requireResource("jira.webresources:ajax-preview", null)
-$webResourceManager.requireResource("org.everit.jira.timetracker.plugin:chart_report-resources")
+$webResourceManager.requireResource("org.everit.jira.timetracker.plugin:table_report-resources")
 <body>
 #set($rendererType = "atlassian-wiki-renderer")
-#if($currentCalendar.useISO8601 == true)
-	#set ($firstDay = 1)
-#else
-	#set ($firstDay = 0)
-#end
-
-<script type="text/javascript" src="https://www.google.com/jsapi"></script>
+	#if($currentCalendar.useISO8601 == true)
+		#set ($firstDay = 1)
+	#else
+		#set ($firstDay = 0)
+	#end
+	
 <script type="text/javascript">
-
 	window.everit = window.everit || {};
 	everit.jttp = everit.jttp || {};
 	everit.jttp.report_common_scripts = everit.jttp.report_common_scripts || {};
@@ -47,48 +44,30 @@ $webResourceManager.requireResource("org.everit.jira.timetracker.plugin:chart_re
 				firstDay : ${firstDay},
 				useISO8601 : ${currentCalendar.useISO8601},
 				dateFromFormated : '${dateFromFormated}',
-				dateFromFormated : '${dateToFormated}'
+				dateToFormated : '${dateToFormated}'
 			}
 		);
 	})(everit.jttp.report_common_scripts, AJS.$);
-	
-	
-	everit.jttp.chart_report = everit.jttp.chart_report || {};
-	#disable_html_escaping()
-	(function(jttp, jQuery) {
-		jttp.options = jQuery.extend( 
-			{},
-			jttp.options,
-			{
-				chartDataList : ${chartDataList}
-			}
-		);
-	})(everit.jttp.chart_report, AJS.$);
-	#enable_html_escaping()
-
 </script>
-
+	
 <div id="plugin_head" class="item-header">
 	<div class="logo-img">
 		<a href="http://www.everit.biz/" target="_blank">
-			<img src="$contextPath/download/resources/org.everit.jira.timetracker.plugin:chart_report-resources/everit_tt_logo.png" alt="Everit logo" height="42" width="42" style:"float: right">
+			<img src="$contextPath/download/resources/org.everit.jira.timetracker.plugin:table_report-resources/everit_tt_logo.png" alt="Everit logo" height="42" width="42" style:"float: right">
 		</a>
 	</div>
 	<h1 class="item-summary">Timetracker</h1>
 </div>
 <div id="main-content" class="content rounded">
-	<form name="chartReportForm"  method="get" action="JiraTimetrackerChartWebAction.jspa" />   
+	<form name="chartReportForm"  method="get" action="JiraTimetrackerTableWebAction.jspa" />   
 		<div class="inside-content">
 			##TAB START
 			<div class="tabwrap tabs2" id="tabs-jttp">
 	    		<ul class="tabs horizontal">
 	        		<li><a href="JiraTimetrackerWebAction!default.jspa"><strong>Timetracker</strong></a></li>
 	        		<li><a href="JiraTimetrackerWorklogsWebAction!default.jspa"><strong>Missing Days Report</strong></a></li>
-	        		<li class="active">
-	            		<a href="JiraTimetrackerChartWebAction!default.jspa"><strong>Chart Report</strong></a>
-	        		</li>
-					<li><a href="JiraTimetrackerTableWebAction!default.jspa"><strong>Table Report</strong></a></li>
-	      </ul>
+	        		<li><a href="JiraTimetrackerChartWebAction!default.jspa"><strong>Chart Report</strong></a></li>
+	        		<li class="active"><a href="JiraTimetrackerTableWebAction!default.jspa"><strong>Table Report</strong></a></li>
 	    		</ul>
 	    	</div>
 			##TAB END
@@ -115,7 +94,7 @@ $webResourceManager.requireResource("org.everit.jira.timetracker.plugin:chart_re
 								$i18n.getText("plugin.to"): 
 								</td>
 								<td>
-								<input type="text" id="dateTo" name="dateTo" tabindex="1" onkeydown="return everit.jttp.report_common_scripts.checkToEnter(event);" value="$dateToFormated"/>
+								<input type="text" id="dateTo" name="dateTo" tabindex="1" onkeydown="return everit.jttp.report_common_scripts.checkEnter(event);" value="$dateToFormated"/>
 								<img id="date_trigger_to" src="$contextPath/images/icons/cal.gif" width="16" height="16" border="0"
 									alt="Pick a date" style="vertical-align: sub;" title="$i18n.getText('plugin.choose.date')"/>
 							</td>
@@ -142,44 +121,21 @@ $webResourceManager.requireResource("org.everit.jira.timetracker.plugin:chart_re
 						<tr>
 							<td>
 								<input id="search"  tabindex="3" type="submit" name="search" value="$i18n.getText('plugin.search')" class="button" 
-								  onclick="return everit.jttp.report_common_scripts.jttpSearchOnClick('Chart');"/>
+								  onclick="return everit.jttp.report_common_scripts.jttpSearchOnClick('Table');"/>
 							</td>
 						</tr>
 					</tbody>
 				</table>
 			</div>
-			#if($chartDataList && $chartDataList.size() == 0)
-				<div id="message" class="message" >
-							 			$i18n.getText("plugin.no.worklogs")
-				</div>
-				<div id="warn_message" class="warn-message" >
-							 			$i18n.getText("plugin.worklog.permission.warning")
-				</div>
-			#elseif($chartDataList)
-				<div id="dashboard">
-					<table>
-						<tbody>
-							<tr>
-								<td>
-						    	<div id="control_div"></div>
-						    	</td>
-						    	<td>
-						    	<div id="chart_div"></div>
-								</td>
-							</tr>
-						</tbody>
-					</table>
-				</div>
-			#end
+			#parse("templates/timetracker/table_report_list.vm")
 		</div>
 	</form>
 	<div id="feedbackfooter" class="feedback_footer">
-		#parse("templates/feedback_chart.vm")
+		#parse("templates/feedback/feedback_table.vm")
 	</div>
 </div>
-
 #if($analyticsCheck)
-	#parse("templates/analytics.vm")
+	#parse("templates/common/analytics.vm")
 #end
 </body>
 </html>
