+++ src/main/resources/templates/timetracker/table_report_list.vm
